app/fastapi/requirements.txt
httpx==0.24.1
fastapi==0.103.1
uvicorn==0.23.2
prometheus-client==0.20.0
opentelemetry-distro==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-httpx==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-jaeger==1.19.0
pyroscope-io==0.8.6
